Extend time counters from 23 bits (float) to 31 bits (int).
In most places, they were stored as doubles (mantissa of 52 bits), but
often passed as floats (mantissa 23 bits).  Such bugs were present in
stabwound's 92M turns game but in that version of Crawl we didn't use
absolute counters for much and turn count was (and is) an int so nothing
was visibly broken.  In current Crawl, it would either crash or trigger
an assert due to counters not adding up.

This commit drops all abuses of double to store integer values.

+++ crawl-ref/source/externs.h
@@ -618,12 +618,12 @@ public:
     long turns_resting;
     long turns_other;
 
-    double elapsed_total;
-    double elapsed_explore;
-    double elapsed_travel;
-    double elapsed_interlevel;
-    double elapsed_resting;
-    double elapsed_other;
+    long elapsed_total;
+    long elapsed_explore;
+    long elapsed_travel;
+    long elapsed_interlevel;
+    long elapsed_resting;
+    long elapsed_other;
 
 public:
     PlaceInfo();
